requirements.txt
PyMuPDF==1.23.8
python-docx==1.1.0
